First part of migration to Angular2 2.0.0
The previous code was developped against Angular2 2.0.0-RC1 which
was already quite outdated.
Since then, GA went out.

This commit is the first attempt at migrating the code to GA.
It is still not entirely stable: typeahead doesn't work (needs an
Observable as a source now), environment.ts isn't overwritten by
the build as expected before, etc...

But this is quite a large first step.

Note: you may need to uninstall the global angular-cli and reinstall
it in its latest version to be able to ng serve, etc...
> npm uninstall -g angular-cli
> npm cache clean
> npm install -g angular-cli@latest